requirements.txt
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.2.0
coverage==6.3.1
crayons==0.4.0
cryptography==35.0.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.12
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-haystack==3.1.1
django-import-export==2.7.1
django-jazzmin==2.4.9
django-pure-pagination==0.3.0
django-qiniu-storage==2.3.1
django-recaptcha==3.0.0
django-redis==5.2.0
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.1.0
docopt==0.6.2
docutils==0.18
et-xmlfile==1.1.0
executing==0.8.2
fabric==2.6.0
fake-useragent==0.1.11
Faker==12.3.0
fett==0.3.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
imageio==2.10.4
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.0
invoke==1.6.0
ipython==8.0.1
jedi==0.18.1
jieba==0.42.1
Markdown==3.3.6
MarkupPy==1.14
matplotlib-inline==0.1.3
moviepy==1.0.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysql-to-sqlite3==1.4.10
mysqlclient==2.1.0
networkx==2.6.3
numpy==1.22.2
odfpy==1.4.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
paramiko==2.8.0
parso==0.8.3
pathlib2==2.3.6
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipdeptree==2.2.1
platformdirs==2.5.0
proglog==0.1.9
prompt-toolkit==3.0.23
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydub==0.25.1
Pygments==2.10.0
pymemcache==3.5.0
pymongo==3.12.1
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jsonrpc-server==0.3.4
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2021.3
PyYAML==5.4.1
qcloudsms-py==0.1.4
qiniu==7.5.0
redis==4.1.3
requests==2.27.1
selenium==4.1.0
sh==1.14.2
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
snooty-lextudio==1.11.5
sortedcontainers==2.4.0
SpeechRecognition==3.8.1
sqlite3-to-mysql==1.4.9
sqlparse==0.4.2
stack-data==0.1.4
tablib==3.2.0
tabulate==0.8.9
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
ua-parser==0.10.0
ujson==1.35
Unidecode==1.3.2
urllib3==1.26.7
user-agents==2.2.0
watchdog==1.0.2
wcwidth==0.2.5
webdriver-manager==3.5.2
Whoosh==2.7.4
wrapt==1.13.3
wsproto==1.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
